// src/API/AutoShopApi/models/AllCarsDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AutoShopAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CarDTO } from './CarDTO';
import {
    CarDTOFromJSON,
    CarDTOFromJSONTyped,
    CarDTOToJSON,
    CarDTOToJSONTyped,
} from './CarDTO';

/**
 * 
 * @export
 * @interface AllCarsDTO
 */
export interface AllCarsDTO {
    /**
     * 
     * @type {Array<CarDTO>}
     * @memberof AllCarsDTO
     */
    cars?: Array<CarDTO> | null;
    /**
     * 
     * @type {number}
     * @memberof AllCarsDTO
     */
    totalCount?: number;
}

/**
 * Check if a given object implements the AllCarsDTO interface.
 */
export function instanceOfAllCarsDTO(value: object): value is AllCarsDTO {
    return true;
}

export function AllCarsDTOFromJSON(json: any): AllCarsDTO {
    return AllCarsDTOFromJSONTyped(json, false);
}

export function AllCarsDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): AllCarsDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'cars': json['cars'] == null ? undefined : ((json['cars'] as Array<any>).map(CarDTOFromJSON)),
        'totalCount': json['totalCount'] == null ? undefined : json['totalCount'],
    };
}

export function AllCarsDTOToJSON(json: any): AllCarsDTO {
    return AllCarsDTOToJSONTyped(json, false);
}

export function AllCarsDTOToJSONTyped(value?: AllCarsDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'cars': value['cars'] == null ? undefined : ((value['cars'] as Array<any>).map(CarDTOToJSON)),
        'totalCount': value['totalCount'],
    };
}